Refactor MsalServiceExceptionFactory
