+++ src/client/Microsoft.Identity.Client.Services/MsalServiceExceptionFactory.cs
@@ -5,16 +5,14 @@ using System;
 using System.Collections.Generic;
 using Microsoft.Identity.Client.Http;
 using Microsoft.Identity.Client.Internal;
-using Microsoft.Identity.Client.Internal.Broker;
 using Microsoft.Identity.Client.OAuth2;
-using Microsoft.Identity.Client.OAuth2.Throttling;
 using Microsoft.Identity.Client.Utils;
 
 namespace Microsoft.Identity.Client
 {
     internal class MsalServiceExceptionFactory
     {
-        static ISet<string> s_nonUiSubErrors = new HashSet<string>(
+        static readonly ISet<string> s_nonUiSubErrors = new HashSet<string>(
             new[] { MsalError.ClientMismatch, MsalError.ProtectionPolicyRequired },
             StringComparer.OrdinalIgnoreCase);
 
@@ -47,10 +45,7 @@ namespace Microsoft.Identity.Client
                     innerException);
             }
 
-            if (ex == null)
-            {
-                ex = new MsalServiceException(errorCode, errorMessage, innerException);
-            }
+            ex ??= new MsalServiceException(errorCode, errorMessage, innerException);
 
             SetHttpExceptionData(ex, httpResponse);
 
@@ -65,53 +60,6 @@ namespace Microsoft.Identity.Client
         {
             return oAuth2Response.ErrorDescription != null &&
                oAuth2Response.ErrorDescription.StartsWith(Constants.AadThrottledErrorCode);
-        }
-
-        internal static MsalServiceException FromBrokerResponse(
-          MsalTokenResponse msalTokenResponse,
-          string errorMessage)
-        {
-            string errorCode = msalTokenResponse.Error;
-            string correlationId = msalTokenResponse.CorrelationId;
-            string subErrorCode = string.IsNullOrEmpty(msalTokenResponse.SubError)?
-                                                                     MsalError.UnknownBrokerError : msalTokenResponse.SubError;
-            HttpResponse brokerHttpResponse = msalTokenResponse.HttpResponse;
-            MsalServiceException ex = null;
-
-            if (IsAppProtectionPolicyRequired(errorCode, subErrorCode))
-            {
-                ex = new IntuneAppProtectionPolicyRequiredException(errorCode, subErrorCode)
-                {
-                    Upn = msalTokenResponse.Upn,
-                    AuthorityUrl = msalTokenResponse.AuthorityUrl,
-                    TenantId = msalTokenResponse.TenantId,
-                    AccountUserId = msalTokenResponse.AccountUserId,
-                };
-            }
-
-            if (IsInvalidGrant(errorCode, subErrorCode) || IsInteractionRequired(errorCode))
-            {
-                ex = new MsalUiRequiredException(errorCode, errorMessage);
-            }
-
-            if (string.Equals(errorCode, MsalError.InvalidClient, StringComparison.OrdinalIgnoreCase))
-            {
-                ex = new MsalServiceException(
-                    MsalError.InvalidClient,
-                    MsalErrorMessage.InvalidClient + " Original exception: " + errorMessage);
-            }
-
-            if (ex == null)
-            {
-                ex = new MsalServiceException(errorCode, errorMessage);
-            }
-
-            SetHttpExceptionData(ex, brokerHttpResponse);
-
-            ex.CorrelationId = correlationId;
-            ex.SubError = subErrorCode;
-
-            return ex;
         }
 
         internal static MsalServiceException FromImdsResponse(
@@ -134,35 +82,22 @@ namespace Microsoft.Identity.Client
             return new MsalThrottledServiceException(ex);
         }
 
-        private static void SetHttpExceptionData(MsalServiceException ex, HttpResponse httpResponse)
+        protected static void SetHttpExceptionData(MsalServiceException ex, HttpResponse httpResponse)
         {
             ex.ResponseBody = httpResponse?.Body;
             ex.StatusCode = httpResponse != null ? (int)httpResponse.StatusCode : 0;
             ex.Headers = httpResponse?.Headers;
         }
 
-        private static bool IsInteractionRequired(string errorCode)
+        protected static bool IsInteractionRequired(string errorCode)
         {
             return string.Equals(errorCode, MsalError.InteractionRequired, StringComparison.OrdinalIgnoreCase);
         }
 
-        private static bool IsInvalidGrant(string errorCode, string subErrorCode)
+        protected static bool IsInvalidGrant(string errorCode, string subErrorCode)
         {
             return string.Equals(errorCode, MsalError.InvalidGrantError, StringComparison.OrdinalIgnoreCase)
                              && IsInvalidGrantSubError(subErrorCode);
-        }
-
-        private static bool IsAppProtectionPolicyRequired(string errorCode, string subErrorCode)
-        {
-#if iOS
-            return string.Equals(errorCode, BrokerResponseConst.iOSBrokerProtectionPoliciesRequiredErrorCode, StringComparison.OrdinalIgnoreCase)
-                             && string.Equals(subErrorCode, MsalError.ProtectionPolicyRequired, StringComparison.OrdinalIgnoreCase);
-#elif ANDROID
-            return string.Equals(errorCode, BrokerResponseConst.AndroidUnauthorizedClient, StringComparison.OrdinalIgnoreCase)
-                             && string.Equals(subErrorCode, MsalError.ProtectionPolicyRequired, StringComparison.OrdinalIgnoreCase);
-#else
-            return false;
-#endif
         }
 
         private static bool IsInvalidGrantSubError(string subError)
